Nove some tests from test_main to test_preprocessing.
